Update variables parameter to correctly parse query variables
When trying to query against a serverless postgraphql instance, I ran into an issue using query variables. Queries that worked as intended with hardcoded data would fail with no error message when refactored to use variables. I investigated the messages that were being sent and found the issue was with the way it was being parsed in serverless.yml. I updated the parameter to use json instead.