Update Travis config to deploy to AWS and test against AWS endpoint
I refactored the endpoint urls to allow for running against local or aws endpoints, and then added a test-on-aws script to package.json. I then updated the travis.yml to use build steps and include deployment and endpoint testing as part of the CI process.

The current setup is the equivalent of having 1 staging environment. As we move forward I'll make some tweaks to allow for a more complete test deployment strategy across feature/staging/master/prod/etc branches.