Merge in Frontend and update Travis deployment
I decided to combine the frontend and backend into a single repo to allow for easier testing, development, and deployments. I moved backend into it's own folder and updated the travis scripts to work. Then I added a frontend folder and moved the frontend into there, and updated travis to run the local frontend tests and deploy the frontend to an s3 bucket. I also updated the frontend to work with all the backend schema changes.