Update Travis scripts to not fail when skipping deployment
There are going to be branches that we don't want to set up AWS deployments for, but we still want to be able to use Travis to ensure the tests pass. I was hoping there would be a way to actually skip the steps conditionally, but that has not been implemented as a part of Travis build stages as of now. In the meantime, I updated the deployment related stages to check for AWS credentials and exit as passing if they aren't set.