Add coordinates to Crossings
We know we'll need each crossing to be tied to a geographical point. As a first step, I made a new column in the table with type= text for each Crossing instance as a required field. I made sure the tests still pass and added a test to make sure creating a Crossing fails if there is no coordinate value. Next, we need to integrate PostGIS into postgraphql and change the value type to a PostGIS coordinate point (or whatever geometry type). ⚔️