Add search user function and tests for search/filter user functionality
In order to implement filtering for the user table on the admin dash, we will need to send filter data into the GraphQL query. Before implementing this on the frontend, I decided to write a few backend tests to ensure the filtering is working correctly.

I also implemented a search function in the schema to search users by firstName, lastName, and community name. This is a very basic search, but satisfies proof of concept requirements. As we groom out the requirements around search, we should write more tests to ensure all search queries function as intended. When writing the new tests, we should create users with unique names/community names to ensure our results are accurate in the case of any async db access.