Add coordinates to Crossing in Front End
The last piece to resolve issue #17 was to display the coordinate point in the React front end. In order to do this, we wrote a new sql function that transforms the coordinate point into a string using `ST_AsLatLonText`. In the future, we may find a new to do this transform into a geojson point as well.