Add viewports to communities and locate user control
In order to show the correct area of the map for any given community, I added viewports to each community in the DB. I also updated the addCrossing function, adding functionality to recalculate the viewport for the crossing's community to make sure new crossings are included.

I then added an onMapboxStyleLoad function to CrossingMap, this allows us to directly access and manipulate the mapbox-gl map object without using the react-mapbox-gl wrapper. Within this function I added functionality to include a geolocation control.