Update map when crossing is clicked
I updated the map component so when a user clicks a crossing the map centers on it and the icon changes to be a selected icon. In order to do this I utilized the mapbox filter functionality, and added a layer for selected crossings. During the styling story, we should implement more layers so we can have different icons for selected crossings based on their status.