Fix status updates
When testing out the crossing list item I found that updating the status of a crossing was setting the latest status of all crossings. This was not being caught by our backend tests, so I updated the status update tests to check that only the desired crossing was being updated. I then updated the schema logic and fixed a minor error to make the tests pass.