Add totals to crossing status toggle in crossing list header (#56)
In order to add the totals to the header, I decided to use a graphql query that gets a totalCount for allCrossings using a condition based on current status. This wasn't possible using the latestStatusUpdateId, as we can't do multi-step conditionals with GraphQL. I updated the backend so we are now storing the latest status id (open/closed/caution/longterm) as well as the latest status update id (with more specific info) on each crossing. I updated the related queries and tests.

I also updated the CSS around the filter checkboxes to make the min width bigger, a possible future improvement would be to allow for smaller widths using multi-line text, but that felt out of scope for what I was working on here.