Edit crossing details (#66)
When working on the edit crossing functionality for the crossing detail page I found that we didn't have an edit crossing function on the backend. I updated the schema to add the function, then added some tests to the backend to ensure the graphql mutation works correctly. I then hooked the crossing detail component up to the mutation.