Display Crossing History
First, we updated the status history query to get all the data. We refactored the DateTime component for shared use on update history list and we introduce PropTypes to documented and predictable typing. Unify UI patterns between Crossing List and Status History

In this mix, we did some restructuring of sass mixins and globals. Simi figured out how to banish relative paths.